added documentation for the logger package
