+++ pkg/logger/README.md
@@ -1,0 +1,231 @@
+# logger
+
+Package logger provides a multi faceted logger that
+can be used to log to the console, a file, or a channel.
+
+<!-- gomarkdoc:embed:start -->
+
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# logger
+
+```go
+import "github.com/conneroisu/steroscopic-hardware/pkg/logger"
+```
+
+Package logger provides a multi faceted logger that can be used to log to the console, a file, or a channel.
+
+It's intended to be used as a default logger for the application.
+
+Allowing for the logging of console messages both to the console and to the browser.
+
+## Index
+
+- [func SetLogger\(\) chan LogEntry](<#SetLogger>)
+- [type ChannelHandler](<#ChannelHandler>)
+  - [func NewChannelHandler\(logChan chan LogEntry, level slog.Level\) \*ChannelHandler](<#NewChannelHandler>)
+  - [func \(h \*ChannelHandler\) Enabled\(\_ context.Context, level slog.Level\) bool](<#ChannelHandler.Enabled>)
+  - [func \(h \*ChannelHandler\) Handle\(ctx context.Context, r slog.Record\) error](<#ChannelHandler.Handle>)
+  - [func \(h \*ChannelHandler\) SetLevel\(level slog.Level\)](<#ChannelHandler.SetLevel>)
+  - [func \(h \*ChannelHandler\) WithAttrs\(attrs \[\]slog.Attr\) slog.Handler](<#ChannelHandler.WithAttrs>)
+  - [func \(h \*ChannelHandler\) WithGroup\(name string\) slog.Handler](<#ChannelHandler.WithGroup>)
+- [type LogEntry](<#LogEntry>)
+- [type Logger](<#Logger>)
+  - [func NewLogger\(\) Logger](<#NewLogger>)
+  - [func \(l \*Logger\) Channel\(\) chan LogEntry](<#Logger.Channel>)
+- [type MultiHandler](<#MultiHandler>)
+  - [func NewMultiHandler\(handlers ...slog.Handler\) \*MultiHandler](<#NewMultiHandler>)
+  - [func \(h \*MultiHandler\) Enabled\(ctx context.Context, level slog.Level\) bool](<#MultiHandler.Enabled>)
+  - [func \(h \*MultiHandler\) Handle\(ctx context.Context, r slog.Record\) error](<#MultiHandler.Handle>)
+  - [func \(h \*MultiHandler\) SetLevel\(level slog.Level\)](<#MultiHandler.SetLevel>)
+  - [func \(h \*MultiHandler\) WithAttrs\(attrs \[\]slog.Attr\) slog.Handler](<#MultiHandler.WithAttrs>)
+  - [func \(h \*MultiHandler\) WithGroup\(name string\) slog.Handler](<#MultiHandler.WithGroup>)
+
+
+<a name="SetLogger"></a>
+## func [SetLogger](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/logger.go#L22>)
+
+```go
+func SetLogger() chan LogEntry
+```
+
+SetLogger sets the default logger to a slog.Logger.
+
+<a name="ChannelHandler"></a>
+## type [ChannelHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L10-L16>)
+
+ChannelHandler implements slog.Handler and sends logs to a channel
+
+```go
+type ChannelHandler struct {
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NewChannelHandler"></a>
+### func [NewChannelHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L19>)
+
+```go
+func NewChannelHandler(logChan chan LogEntry, level slog.Level) *ChannelHandler
+```
+
+NewChannelHandler creates a new ChannelHandler
+
+<a name="ChannelHandler.Enabled"></a>
+### func \(\*ChannelHandler\) [Enabled](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L27>)
+
+```go
+func (h *ChannelHandler) Enabled(_ context.Context, level slog.Level) bool
+```
+
+Enabled implements slog.Handler.Enabled
+
+<a name="ChannelHandler.Handle"></a>
+### func \(\*ChannelHandler\) [Handle](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L32>)
+
+```go
+func (h *ChannelHandler) Handle(ctx context.Context, r slog.Record) error
+```
+
+Handle implements slog.Handler.Handle
+
+<a name="ChannelHandler.SetLevel"></a>
+### func \(\*ChannelHandler\) [SetLevel](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L99>)
+
+```go
+func (h *ChannelHandler) SetLevel(level slog.Level)
+```
+
+SetLevel sets the minimum level for logging
+
+<a name="ChannelHandler.WithAttrs"></a>
+### func \(\*ChannelHandler\) [WithAttrs](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L64>)
+
+```go
+func (h *ChannelHandler) WithAttrs(attrs []slog.Attr) slog.Handler
+```
+
+WithAttrs implements slog.Handler.WithAttrs
+
+<a name="ChannelHandler.WithGroup"></a>
+### func \(\*ChannelHandler\) [WithGroup](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/channel.go#L82>)
+
+```go
+func (h *ChannelHandler) WithGroup(name string) slog.Handler
+```
+
+WithGroup implements slog.Handler.WithGroup
+
+<a name="LogEntry"></a>
+## type [LogEntry](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/logger.go#L48-L53>)
+
+LogEntry represents a structured log entry
+
+```go
+type LogEntry struct {
+    Level   slog.Level
+    Time    time.Time
+    Message string
+    Attrs   []slog.Attr
+}
+```
+
+<a name="Logger"></a>
+## type [Logger](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/logger.go#L11-L14>)
+
+Logger is a slog.Logger that sends logs to a channel and also to the console.
+
+```go
+type Logger struct {
+    *slog.Logger
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NewLogger"></a>
+### func [NewLogger](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/logger.go#L34>)
+
+```go
+func NewLogger() Logger
+```
+
+NewLogger creates a new Logger.
+
+<a name="Logger.Channel"></a>
+### func \(\*Logger\) [Channel](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/logger.go#L17>)
+
+```go
+func (l *Logger) Channel() chan LogEntry
+```
+
+Channel returns the channel to which logs are sent to the browser.
+
+<a name="MultiHandler"></a>
+## type [MultiHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L10-L13>)
+
+MultiHandler combines multiple slog.Handler instances
+
+```go
+type MultiHandler struct {
+    // contains filtered or unexported fields
+}
+```
+
+<a name="NewMultiHandler"></a>
+### func [NewMultiHandler](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L16>)
+
+```go
+func NewMultiHandler(handlers ...slog.Handler) *MultiHandler
+```
+
+NewMultiHandler creates a new MultiHandler
+
+<a name="MultiHandler.Enabled"></a>
+### func \(\*MultiHandler\) [Enabled](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L21>)
+
+```go
+func (h *MultiHandler) Enabled(ctx context.Context, level slog.Level) bool
+```
+
+Enabled implements slog.Handler.Enabled
+
+<a name="MultiHandler.Handle"></a>
+### func \(\*MultiHandler\) [Handle](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L31>)
+
+```go
+func (h *MultiHandler) Handle(ctx context.Context, r slog.Record) error
+```
+
+Handle implements slog.Handler.Handle
+
+<a name="MultiHandler.SetLevel"></a>
+### func \(\*MultiHandler\) [SetLevel](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L58>)
+
+```go
+func (h *MultiHandler) SetLevel(level slog.Level)
+```
+
+SetLevel sets the minimum level for all handlers that support it
+
+<a name="MultiHandler.WithAttrs"></a>
+### func \(\*MultiHandler\) [WithAttrs](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L40>)
+
+```go
+func (h *MultiHandler) WithAttrs(attrs []slog.Attr) slog.Handler
+```
+
+WithAttrs implements slog.Handler.WithAttrs
+
+<a name="MultiHandler.WithGroup"></a>
+### func \(\*MultiHandler\) [WithGroup](<https://github.com/conneroisu/steroscopic-hardware/blob/main/pkg/logger/multi.go#L49>)
+
+```go
+func (h *MultiHandler) WithGroup(name string) slog.Handler
+```
+
+WithGroup implements slog.Handler.WithGroup
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
